+++ client/src/App.js
@@ -1,11 +1,23 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Home from './pages/Home';
 import Dashboard from './pages/Dashboard';
+import withFirebaseAuth from 'react-with-firebase-auth'
+import * as firebase from 'firebase/app';
+import 'firebase/auth';
+import firebaseConfig from './components/Firebase/firebaseConfig';
 
 
+const firebaseApp = firebase.initializeApp(firebaseConfig);
 
-function App() {
+class App extends Component {
+  render() {
+    const {
+      user,
+      signOut,
+      signInWithGoogle,
+  } = this.props;
+
     return (
       <Router>
         <div>
@@ -16,7 +28,16 @@ function App() {
           </div>
       </Router>
     );
+  }
 }
 
 
-export default App;
+const firebaseAppAuth = firebaseApp.auth();
+const providers = {
+  googleProvider: new firebase.auth.GoogleAuthProvider(),
+};
+
+export default withFirebaseAuth({
+    providers,
+    firebaseAppAuth,
+})(App);
