+++ client/src/components/Firebase/index.js
@@ -1,45 +1,28 @@
 import React, { Component } from 'react';
-import withFirebaseAuth from 'react-with-firebase-auth'
-import * as firebase from 'firebase/app';
-import 'firebase/auth';
-import firebaseConfig from './firebaseConfig';
+import * as firebase from 'firebase';
+import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
 
-const firebaseApp = firebase.initializeApp(firebaseConfig);
 
-class LoginWithGoogle extends Component {
-    render() {
-        const {
-            user,
-            signOut,
-            signInWithGoogle,
-        } = this.props;
-
-    return (
-        <div className="App">
-          <header className="App-header">
-            {
-              user
-                ? <p>Hello, {user.displayName}</p>
-                : <p>Please sign in.</p>
-            }
-  
-            {
-              user
-                ? <button onClick={signOut}>Sign out</button>
-                : <button onClick={signInWithGoogle}>Sign in with Google</button>
-            }
-          </header>
-        </div>
-      );
-    }
-}
-
-const firebaseAppAuth = firebaseApp.auth();
-const providers = {
-  googleProvider: new firebase.auth.GoogleAuthProvider(),
+// This is our firebaseui configuration object
+const uiConfig = {
+    signInFlow: 'popup',
+    signInSuccessUrl: '/dashboard',
+    signInOptions: [
+        firebase.auth.GoogleAuthProvider.PROVIDER_ID
+  ]
 };
 
-export default withFirebaseAuth({
-    providers,
-    firebaseAppAuth,
-  })(LoginWithGoogle);
+class SignIn extends Component {
+      render() {
+        return (
+          <div>
+          <h1>My App</h1>
+            <p>Please sign-in:</p>
+            <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()}/>
+          </div>
+        )
+      }
+}
+
+export default SignIn
+
