+++ client/src/pages/Home.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Container, Row, Col } from "../components/Grid/index.js";
 import "./Home.css";
-import LoginWithGoogle from "../components/Firebase/index"
+import SignIn from "../components/Firebase/index"
 
 class Home extends Component {
     render() {
@@ -13,7 +13,7 @@ class Home extends Component {
                     </Col>
                 </Row>
                 <Row>
-                    < LoginWithGoogle />
+                    <SignIn />
                 </Row>
 
             </Container>
