+++ package.json
@@ -25,6 +25,7 @@
     "firebaseui": "^4.3.0",
     "if-env": "^1.0.4",
     "mongoose": "^5.7.11",
+    "react-firebaseui": "^4.0.0",
     "react-router-dom": "^5.1.2"
   }
 }
